+++ packages/stylelint-config-wordpress/__tests__/properties.js
@@ -6,6 +6,7 @@ import test from "ava"
 const validCss = (`
 #selector-1 {
 	background: #fff;
+	display: block;
 	margin: 0;
 	margin-left: 20px;
 }
@@ -14,6 +15,7 @@ const validCss = (`
 const invalidCss = (`
 #selector-1 {
 	background:#FFFFFF;
+	display: BLOCK;
 	margin-left: 20PX;
 	margin: 0;
 }
@@ -42,11 +44,12 @@ test("There are warnings with invalid properties CSS", t => {
     const { errored, results } = data
     const { warnings } = results[0]
     t.truthy(errored, "errored")
-    t.is(warnings.length, 5, "flags eight warnings")
+    t.is(warnings.length, 6, "flags six warnings")
     t.is(warnings[0].text, "Expected \"#FFFFFF\" to be \"#ffffff\" (color-hex-case)", "correct warning text")
     t.is(warnings[1].text, "Expected \"#FFFFFF\" to be \"#FFF\" (color-hex-length)", "correct warning text")
     t.is(warnings[2].text, "Unexpected shorthand \"margin\" after \"margin-left\" (declaration-block-no-shorthand-property-overrides)", "correct warning text")
     t.is(warnings[3].text, "Expected single space after \":\" with a single-line declaration (declaration-colon-space-after)", "correct warning text")
     t.is(warnings[4].text, "Expected \"PX\" to be \"px\" (unit-case)", "correct warning text")
+    t.is(warnings[5].text, "Expected \"BLOCK\" to be \"block\" (value-keyword-case)", "correct warning text")
   })
 })