+++ packages/stylelint-config-wordpress/index.js
@@ -86,6 +86,7 @@ module.exports = {
     "string-quotes": "double",
     "unit-case": "lower",
     "unit-no-unknown": true,
+    "value-keyword-case": "lower",
     "value-list-comma-newline-after": "always-multi-line",
     "value-list-comma-space-after": "always-single-line",
     "value-list-comma-space-before": "never",