+++ packages/stylelint-config-wordpress/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Removed: `stylelint < 6.6.0` compatibility.
 * Removed: `number-zero-length-no-unit` rule.
 * Added: `length-zero-no-unit` rule.
+* Added: `value-keyword-case` rule.
 
 # 7.1.1